Using and outputting dynamic values.
App update. Using and outputting dynamic values.
