+++ 01-starting-project/src/App.jsx
@@ -1,4 +1,4 @@
-import Header from "../Header";
+import Header from "./components/Header";
 
 const App = () => {
   return (
