+++ 01-starting-project/src/components/Header.jsx
@@ -1,0 +1,22 @@
+const reactDiscriptions = ['Fundamental', 'Crucial', 'Core'];
+
+function genRandomInt(max) {
+    return Math.floor(Math.random() * (max + 1));
+}
+
+const Header = () => {
+    const description = reactDiscriptions[genRandomInt(2)];
+
+    return (
+      <header>
+        <img src="src/assets/react-core-concepts.png" alt="" />
+        <h1>React Essentials</h1>
+        <p>
+          {description} React concepts you will need for almost any app you are
+          going to build!
+        </p>
+      </header>
+    )
+}
+
+export default Header;
